working symbolication scripts.
How to use:
Run the sample_once example. This will print each thread's stack, then
the proc maps entry.

Run the relative_addr.py script with the base address of the map entry
for sample_once, and pass as stdin the set of `IP: 0x...` lines for a
thread. This will print out the relative addresses.

Now run the lookup_symbol example with the path to the sym file and the
relative address of the symbol to lookup. If one is found, it will be
printed.

If sample_once is changed, a new sym file can be generated by running

    dump_syms -i <path to binary>

This will get the build ID required to build up the file name.

    mkdir -p symbols/<binary name>/<build ID>
    dump_syms <path to binary> > symbols/<binary name>/<build ID>/<binary name>.sym