Ton of improvements around output
Added several binaries. The new workflow is something like:

1. Use sample_once to generate a vignette unresolved sample.
sample_once was changed to sample multiple times.

2. Use the generate_symbols binary with a binary and symbols location
This will send the binary through dump_syms and write the sym file to
the write build ID based location under the symbol root.

TODO: Do this for the dlls the binary loads.

3. Use the resolve binary to convert the profile to a resolved profile
This uses the symbolic crate(s) to use the sym file to lookup symbols
using the IP and write out a resolved profile with function, file and
line info.

4. Use the speed binary to convert the resolved profile to a SpeedScope
profile
SpeedScope.app is a nice web based viewer like gecko's, but with a
simpler and well specced profile format. Copied the rust code from rbspy
with some modifications.
This will convert to a speed file, which can be loaded in the browser
and looks great!
Remember that the speedscope frame ordering is the reverse of ours.

Finally, a lot of cleanup is possible, which should probably be the next
area of focus
1. use clap for all the binaries
2. move VecHashMap to a useful place
3. Clean up SymCacheCache with some better understanding of lifetimes
4. Make speed.rs a proper interface instead of exposing Frame publicly